Add a multiplier to the system clock and divisors to the clocked components
This should allow components to be run at different rates, eg in the master system the CPU runs at the master clock / 3 and the vdp runs at mater clock / 2.

The CPU "produces" the cycles which will be multiplied to generate the master clock. The components will then use the master clock divided down to generate their update clocks.